clean code & fix
