+++ src/pages/api/project/update.tsx
@@ -18,7 +18,6 @@ export default async function(req: NextApiRequest, res: NextApiResponse) {
     res.status(201)
     res.json({ actualProject })
   } catch (e) {
-    console.error(e)
     res.status(500)
     res.json({ error: 'Sorry unable to update project to database' })
   } finally {
